Refactor serializers/projects.py
